+++ trixnity-vodozemac/src/nativeMain/kotlin/net/folivo/trixnity/vodozemac/Library.native.kt
@@ -1,0 +1,9 @@
+@file:OptIn(ExperimentalForeignApi::class)
+
+package net.folivo.trixnity.vodozemac
+
+import kotlinx.cinterop.ExperimentalForeignApi
+
+actual typealias ExternalSymbolName = SymbolName
+
+internal actual val InitHook: () -> Unit = {}