+++ trixnity-vodozemac/src/wasmJsMain/kotlin/net/folivo/trixnity/vodozemac/Library.wasmJs.kt
@@ -1,0 +1,5 @@
+package net.folivo.trixnity.vodozemac
+
+actual typealias Import = JsModule
+
+actual typealias ModuleImport = kotlin.wasm.WasmImport